Attribut "geloest_in" der Entität Ergebnis angepasst - zur Double gemacht

+++ src/main/java/com/hka/exitgame/dto/ErgebnisDto.java
@@ -8,7 +8,7 @@ public record ErgebnisDto(
         UUID spielerId,
         UUID aufgabeId,
         UUID semesterId,
-        Integer geloestIn
+        Double geloestIn
 ) {
 
     public Ergebnis toErgebnis() {
